requirements.txt
streamlit
pdfplumber
pytesseract
Pillow
transformers
torch
protobuf==3.20.*
